pingpong/requirements.txt
python-dotenv>=0.21.0
channels==3.0.4
channels-redis==3.3.1
Django==4.0
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
qrcode==7.4.2
pyotp==2.9.0
pillow==10.4.0
python-dotenv